+++ test/cypress/integration/R4/answerExpression.cy.js
@@ -12,29 +12,29 @@ describe('AnswerExpression ', () => {
 
   it('should update answer list on an item whose type is string', () => {
     cy.byId('q2String/1').click();
-    cy.get('#searchResults li').should('have.length', 3);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 3);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q2String/1').should('have.value', 'one');
   });
 
   it('should update answer list on an item whose type is integer', () => {
     cy.byId('q3Integer/1').click();
-    cy.get('#searchResults li').should('have.length', 3);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 3);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q3Integer/1').should('have.value', '12');
   });
 
   it('should update answer list on an item whose type is date', () => {
     cy.byId('q4Date/1').click();
-    cy.get('#searchResults li').should('have.length', 3);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 3);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q4Date/1').should('have.value', '2022-11-12');
   });
 
   it('should update answer list on an item whose type is time', () => {
     cy.byId('q5Time/1').click();
-    cy.get('#searchResults li').should('have.length', 3);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 3);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q5Time/1').should('have.value', '17:03:07');
   });
 
@@ -42,13 +42,13 @@ describe('AnswerExpression ', () => {
     cy.byId('q1/1').click().type("abc");
 
     cy.byId('q1ListChoice/1').click();
-    cy.get('#searchResults li').should('have.length', 1);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 1);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q1ListChoice/1').should('have.value', 'abc');
 
     cy.byId('q1ListString/1').click();
-    cy.get('#searchResults li').should('have.length', 1);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 1);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q1ListString/1').should('have.value', 'abc');
 
     // added another q1
@@ -57,15 +57,15 @@ describe('AnswerExpression ', () => {
     // the existing answer lists on 2 items disappeared (expected behavior)
     cy.byId('q1ListChoice/1').should('have.value', '');
     cy.byId('q1ListString/1').should('have.value', '');
-    
+
     cy.byId('q1ListChoice/1').click();
-    cy.get('#searchResults li').should('have.length', 2);
-    cy.get('#searchResults li:first-child').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 2);
+    cy.get('#lhc-tools-searchResults li:first-child').click();
     cy.byId('q1ListChoice/1').should('have.value', 'abc');
 
     cy.byId('q1ListString/1').click();
-    cy.get('#searchResults li').should('have.length', 2);
-    cy.get('#searchResults li:nth-child(2)').click();
+    cy.get('#lhc-tools-searchResults li').should('have.length', 2);
+    cy.get('#lhc-tools-searchResults li:nth-child(2)').click();
     cy.byId('q1ListString/1').should('have.value', 'def');
 
   });