+++ test/cypress/integration/add-remove-repeating-row.cy.ts
@@ -121,3 +121,31 @@ describe('repeating group with answerValueSet items', () => {
   });
 });
 
+describe('repeating group with tooltip to show on empty items', () => {
+  const tp: any = new TestPage();
+
+  it('should display tooltips correctly when adding a repeating group', () => {
+    tp.openBaseTestPage();
+    tp.loadFromTestData('test-tooltip-on-repeat-item.json', 'R4');
+    const allergy1 = '/allergies/allergy_name/1/1';
+    const allergy2 = '/allergies/allergy_name/2/1';
+    const allergy3 = '/allergies/allergy_name/3/1';
+    const addButton = 'add-/allergies/1';
+    const addTooltip = 'add-content-/allergies/1';
+    cy.byId(allergy1).click();
+    cy.byId(allergy1).type('{downarrow}').type('{enter}');
+    cy.byId(allergy1).should('have.value', 'Chocolate');
+    cy.byId(addButton).click();
+    cy.byId(allergy2).should('be.visible');
+    cy.byId(addTooltip).should('not.exist');
+    cy.byId(addButton).click();
+    cy.byId(addTooltip).should('be.visible').should('contain', 'Please enter info in the blank ');
+    cy.byId(allergy2).click();
+    cy.byId(allergy2).type('{downarrow}').type('{downarrow}').type('{enter}');
+    cy.byId(allergy2).should('have.value', 'Crab');
+    cy.byId(addButton).click();
+    cy.byId(allergy3).should('be.visible');
+    cy.byId(addTooltip).should('not.exist');
+  });
+});
+