+++ test/cypress/integration/addFormToPage-error-message.cy.js
@@ -9,6 +9,11 @@ describe('addFormToPage Error Message Test', () => {
     // has an error message
     // This also tests that the url parameter gets URL-encoded.
     cy.get("#loadMsg").contains("Unable to load ValueSet");
+    // An error message is shown under the field for which answerValueSet expansion failed.
+    cy.byId('item-/54126-8/54128-4/1/1')
+      .find('.lhc-item-error')
+      .should('be.visible')
+      .should('have.text', 'Error: Unable to load the answer list for this question.');
   });
 
   it('show an error when a valueset cannot be loaded because of a wrong fhir context', () => {