+++ test/cypress/integration/extract-observations.cy.ts
@@ -88,4 +88,33 @@ describe('Form with extract observation extension', ()=>{
   });
 
 
+  it('should extract Questionnaire.item.code into Observation.code.coding', () => {
+    cy.visit('test/pages/addFormToPageTest.html');
+    util.addFormToPage('extractObsCode-test.R4.json', null, {fhirVersion: 'R4'});
+    cy.window().then((win) => {
+      const bundle = win.LForms.Util.getFormFHIRData("QuestionnaireResponse", "R4", null, {extract: true});
+      expect(bundle.length).to.equal(5);
+      expect(bundle[0].resourceType).to.equal("QuestionnaireResponse");
+      // Should not extract child items if parent item has obsExtract=false.
+      expect(bundle.some(b => b.resourceType === "Observation" && b.code?.text === "intItem0")).to.equal(false);
+      // Should extract child item if it turns extraction back on again with ObsExtract=true.
+      expect(bundle[1].resourceType).to.equal("Observation");
+      expect(bundle[1].code.text).to.equal("intItem1");
+      // Should extract only the code with ObsExtract=true if the extension is on code level.
+      expect(bundle[2].resourceType).to.equal("Observation");
+      expect(bundle[2].code.text).to.equal("blItem3")
+      expect(bundle[2].code.coding.length).to.equal(1);
+      expect(bundle[2].code.coding[0].code).to.equal("code5");
+      // Should extract more than one item.code into Observation.code.
+      expect(bundle[3].resourceType).to.equal("Observation");
+      expect(bundle[3].code.text).to.equal("codingItem2")
+      expect(bundle[3].code.coding.length).to.equal(2);
+      // Should only skip codes with ObsExtract=false if item has ObsExtract=true.
+      expect(bundle[4].resourceType).to.equal("Observation");
+      expect(bundle[4].code.text).to.equal("blItem4")
+      expect(bundle[4].code.coding.length).to.equal(2);
+      expect(bundle[4].code.coding[0].code).to.equal("code6"); // code with ObsExtract=true
+      expect(bundle[4].code.coding[1].code).to.equal("code8"); // code without ObsExtract extension
+    });
+  });
 });